Add limelight id to LimelightConstants structure.
This ID is used in a few places, such as the URL. An ID of 11 corresponds with an IP of 10.49.15.11. While LimelightHelpers doesn't need this, this is useful for port forwarding and also helps us manage the limelights. The kLimelights array in VisionConstants can also act as a list of limelights that is easily accessible.