Change VisionSubsystem to not use getInstance() and instead have a public constructor.
Using getInstance() seems useful: other subsystems can easily access the same version of a subsystem. However, this becomes a problem when one wishes to disable a subsystem when debugging or testing, as one must go through every subsystem and try to find which ones call this getInstance() method. If, instead, we pass all needed subsystems through the constructor, one can quickly see which subsystems use a given subsystem by looking in RobotContainer. This is the design paradigm we will be following this year.