Move OdometrySubsystem logic to vision subsystems.
Standard deviations are calculated by LimelightDevice and are included in the VisionMeasurement structure. The LimelightVisionSubsystem adds pose estimations to the swerve drive. The heuristics for standard deviations are not yet implemented! Also, the getVisionMeasurement method now takes an enum instead of a boolean for which pose estimation method to use for clarity.