Update BlingShow.java to be more memory efficient
Instead of using a text file for the bling, it now uses a file composed of many bytes. Instead of storing the in a short[][], it now stores it in a byte[][]. This should use half the memory and, according to my (limited) testing, use only a third of the storage space.