Combine Marshall's climber code into one mergeable commit with a few notable changes
Magic numbers in the bindings have been moved to enums in ClimberConstants. Those values are all 0.0 as I don't know which direction is positive. These climber bindings are set to the operator dpad buttons, and the manual controls (which used to be operator dpad) are now on debug controller dpad.

Before merging into dynamics, do these things:
- Find which direction is positive and set the speeds for the climber and winch
- Fix DynamicsCommandFactory.gotoClimb() and run it when operator uses climber controls